Another set of refactors
