+++ src/heart/device/rgb_display.py
@@ -205,12 +205,15 @@ class SampleBase(object):
 
 
 class LEDMatrix(Device, SampleBase):
-    def __init__(self, chain_length: int, *args, **kwargs):
+    def __init__(self, layout: Layout, *args, **kwargs):
         super(LEDMatrix, self).__init__(
-            *args, **kwargs, layout=Layout(columns=chain_length, rows=1)
+            *args,
+            **kwargs,
+            layout=layout
         )
+        assert layout.rows == 1, "Maximum 1 row supported at the moment"
 
-        self.chain_length = chain_length
+        self.chain_length = layout.columns
         self.row_size = 64
         self.col_size = 64
