+++ src/heart/display/renderers/image.py
@@ -3,7 +3,7 @@ import pygame
 from heart.assets.loader import Loader
 from heart.device import Orientation
 from heart.display.renderers import BaseRenderer
-from heart.display.renderers.models import KeyFrame
+from heart.display.models import KeyFrame
 from heart.peripheral.manager import PeripheralManager
