+++ src/heart/display/renderers/spritesheet_random.py
@@ -8,7 +8,7 @@ from heart import DeviceDisplayMode
 from heart.assets.loader import Loader
 from heart.device import Orientation
 from heart.display.renderers import BaseRenderer
-from heart.display.renderers.models import KeyFrame
+from heart.display.models import KeyFrame
 from heart.peripheral.manager import PeripheralManager
